requirements.txt
BeautifulSoup==3.2.1
Django==1.6
Fabric==1.8.0
Jinja2==2.7.1
MarkupSafe==0.18
MySQL-python==1.2.5
Pygments==1.6
SQLAlchemy==0.8.4
South==0.8.2
Sphinx==1.2
argparse==1.2.1
beautifulsoup4==4.3.2
boto==2.15.0
chardet==2.2.1
csvkit==0.6.1
dbf==0.95.004
dj-database-url==0.2.2
dj-static==0.0.5
django-appconf==0.6
django-compressor==1.3
django-debug-toolbar==0.10.2
django-tastypie==0.10.0
docutils==0.11
ecdsa==0.10
gunicorn==18.0
ipython==1.1.0
matplotlib==1.3.1
mimeparse==0.1.3
nameparser==0.2.8
nose==1.3.0
numpy==1.8.0
openpyxl==1.7.0
pandas==0.13.1
paramiko==1.12.0
patsy==0.2.1
psycopg2==2.5.1
pycrypto==2.6.1
pyparsing==2.0.1
python-dateutil==2.2
python-memcached==1.53
pytz==2013.9
pyzmq==14.0.1
requests==2.1.0
scikit-learn==0.14.1
scipy==0.13.3
six==1.4.1
sqlparse==0.1.10
static==0.4
statsmodels==0.5.0
tornado==3.1.1
wsgiref==0.1.2
xlrd==0.9.2
yolk==0.4.3
